Ganti nextInt ke nextDouble di matrixmethods

+++ src/matrixmethods.java
@@ -251,7 +251,7 @@ public class matrixmethods {
 
         for (int i = 0; i < m; i++) {
             for (int j = 0; j < n; j++) {
-                newMatrix[i][j] = input.nextInt();
+                newMatrix[i][j] = input.nextDouble();
             }
         }
     }
